Add scheduler for running playbooks (#5)
This is stacked on top of #4. See commits starting with
4faca9b5516ef0375b96b0752fd1deb41520a39f.

This PR adds a simple scheduler / task queue that saves it's data in
Postgres.

At the high level, it works like this:

* "something" periodically calls a Next.js route to give us a clock
tick. This can be a vercel cron that fires every 1 minute, or anything
else. For local development, I added `pnpm run dev-scheduler` that calls
the tick route every 10 seconds.
* when this route is called, we look for jobs that are ready to be run.
If they are, we start up to 20 of them in parallel. If there are more
than 20, we leave the rest for the next run.
* after a run, we compute when we want the next run to be (e.g. based on
the cron expression) and store that in the DB.

This is a simplistic queue but I think will do the job for now. The only
thing that i'm not sure is how we'll call that private tick route when
deploying as OSS. Maybe actual cron in the docker-container.

<img width="1907" alt="Screenshot 2025-02-22 at 23 14 42"
src="https://github.com/user-attachments/assets/66546f99-a9f6-455c-91d9-4ea2e7426799"
/>

---------

Co-authored-by: Alexis Rico <sferadev@gmail.com>