Add schedules history (#16)
Sits on top of #15.

This saves schedules runs in the DB. We automatically remove old runs,
and how much history is kept is a config parameter (defaults to 300).

It also adds a page to view old runs and the ability to load them in
chat.

The PR also updates the READMEs in the repo.

---------

Co-authored-by: Alexis Rico <sferadev@gmail.com>