Adding dockerfile (#32)
Tested and this works to run the next.js app inside docker. 

You can build the image with:

```
docker build -t dbagent .
```

Then run it with something like this:

```
docker run -p 8080:8080 \
   -e DATABASE_URL='postgres://dbagent:changeme@host.docker.internal:5432/dbagent' \
   -e OPENAI_API_KEY='sk-proj-...' \
    -e AUTH_SECRET="secret \
    -e AUTH_TRUST_HOST='localhost:4001' \
    dbagent
```

Just note that the DATABASE_URL needs to use that `host.docker.internal`
host instead of localhost if you are running Postgres in Docker.