Chat persistence
With this change we store the chat and message history in postgres.

The Chat UI supports creation and deletion of chats. Chats also have a
title. The title of the active chat can be modified by clicking on the
title in the chat header.
The UI will not send all messages anymore, but only the most recent
message. We fetch the chat history from the chats memory.

The change introduces a chat memory in lib/ai/memory.ts

The chats are stores in the new chats table. A chat is associated with a
procject and can have an optional connection to act upon.

The chat messages are stored in the messages table.