Split tools
We already have a number of tools. Some tools accessing playbooks, the target DB
and others to access AWS cluster information.

This changes splits the getTools function and groups tools according to
the resources they interact with. This will allow us to selectively
build tools. For example ignore the AWS based tools if the AWS cluster
configuration is missing, or introduce support for other providers like
GCP.