Custom play book (#61)
This change adds the custom Playbook. Users are able to create custom
Playbooks to be ran or scheduled. They have the following features and
qualities:
- Added a custom playbook form that can generate playbook content using
gpt-4o.
- The playbook view table shows both built-in and custom playbooks,
which are distinguished by a type category on the table.
- The custom playbooks have an edit button allowing a user to access and
edit the form to change the playbook description and content, along with
using AI to generate more/new content, if needed, when editing.
- Only custom playbooks can be edited
- The edit form follows a similar style to that of the edit monitor,
with a delete button in the edit form that asks for confirmation before
deleting a custom playbook. If a custom playbook is being used for
monitoring, then an error message appears, and the user is prevented
from deleting the playbook before deleting the monitoring schedule
- The custom playbooks work with both scheduling and the chat(they
appear in the scheduling form and can be accessed by both the chat and
the scheduling form/scheduler)
- The code is split up in a similar way to that of the monitoring
folder, with a separate file under db/ for db interactions and a
separate file under tools/ for tools used by action.ts
- Built-in and Custom playbooks can have their content copied into a new
create custom playbook form; there is now a button to carry this out on
the playbook table page. Each subsequent copy of the same playbook is
identified by a (Copy 1), (Copy 2),...
-The playbook table page only shows the run playbook button; all other
features are in a dropdown menu to the right of the run playbook button