TargetDB tools connection reuse + concurrency limit (#67)
We introduce connection pool for tools accessing the target DB.

The connection pool only creates a connection on demand, but ensures
that the connection is reused between tool calls. In case the model opts
for parallel tool calls the connection pool also ensures that we have a
bounded number of connections to the target DB.

For now the connection pool is only active for a single request. There
is currently no concurrency limit. To limit the number of concurrently
active agent->DB connections one might use pgbouncer or similar tools.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>