Add GCP support in CloudSQL (#77)
This adds the "cloud provider" type at the project level. I think at
project level makes sense in order for us to be able to have different
onboarding guides depending on the Cloud provider. For example, Xata
hosted DBs will be much easier to integrate because we get all the data
at once.

For GCP, specifically, this adds a way to configure the connection, and
tools for retrieving metrics and logs.

I have some TODOs left:

* [x] Create a guide on how to configure the integration, like we have
for GCP.
https://github.com/xataio/agent/wiki/Xata-Agent-%E2%80%90-GCP-CloudSQL-integration-guide
* [x] Simplify the integration by allowing the user to upload the json
file.
* [x] Better test the 'Other' provider, perhaps we can adjust the
playbooks for when we don't have a cloud integration.
* [x] Include the cloud provider type in the system prompt.
* [x] Consider first time experience if it can be improved.

But it's ready for reviews, I can also do this in follow up PRs.