Playground (#56)
Provide an Agent playground based on mastra playground. To run the
playground use `pnpm run playground`.

No mastra libraries are used by the main project. The change refactors
tool setup, so that we can setup and reuse tools from the playground.

The agents are setup with a few evaluators already provided by mastra.
Unfortunately the evalers interfaces only allow us to compare input and
output per request/response. Some evaluators in mastra can evaluate the
response based on user input chat context, but these can not be
registered with the agent.

The use of NextAuth in queryDb did cause some problems to me when
starting the playground. I chose to remove queryDb in favor of passing a
DBAccess instance. With that change we decide in advance the context
(current session user, other user, admin) queries shall be executed
with. The DBAccess method uses a connection pool limiting the number of
concurrent connections requests in case the agent want to call tools in
parallel.