Add reasoning models O1 and O4-mini (#154)
It seems that the reasoning models don't support tools (functions in
OpenAI lingo) with optional parameters. We only had two tools that were
using optional parameters, and we didn't really need that, so for now I
just made the parameters mandatory.

This just adds the models, but we'd want the UI to display the
intermediary reasoning as well, which is not happening currently.