Custom tools via MCP (#153)
**Added a custom MCP server table and MCP view page**
This is a (UI)
tools run in both chat and can be monitored via a playbook
- Allows the user to view mcp servers(only .ts for now) added to the
local mcp-source file along with their file name, enable or disable an
mcp server, and access a dropdown to view server information and
available tools(and their descriptions) for the selected mcp server
- The mcp-view page is accessible via the view details button in the
dropdown
- The mcp-view page allows an mcp server to be deleted from the db and
the local files
- The mcp-view page allows an mcp tool to be run from the mcp-view page
if the server is enabled
- The mcp server table has an add mcp server button that brings the user
to a page with an upload/drag and drop mechanism to add an mcp server
directly to the mcp-source local file. When the server is added, it is
automatically built(using the build:mcp-servers script) to be run as a
.js file by the Agents local mcp client.
- currently mcp servers are run locally using the stdio protocol

**Added a tools page(UI)** 

- allows the user to view all the built in and enabled custom tools and
their details, as well as run them. This includes the mcp server name,
version, enabled/disabled/not run before(needs to be added to db)
status, and Available tools

**Local mcp-source file**
In the local mcp-source file, users can add custom mcp tools directly to
the file. This can also be done via the add mcp servers button
page(disabled for now)

- users can build a .js file from a .ts mcp server file using the
build:mcp-servers, this must be done before a server is available

**Test MCP servers**

- added simple servers with basic tools to test that the mcp server and
its tools are working, eg: (multiplication, add, modulo) tools using two
parameters, a testingTool that outputs the input.

---------

Signed-off-by: Alexis Rico <sferadev@gmail.com>
Co-authored-by: Alexis Rico <sferadev@gmail.com>