fix(docker-compose): restart after hosts restarts (#195)
Currently, agent doesn't restart if the host machine is restarted. This
fixes it. Chose `unless-stopped` over `always`.

| Policy | What happens after you run `docker stop <ctr>` | On next
Docker/host reboot |

|--------|-----------------------------------------------|----------------------------|
| `always` | Container stays stopped **for now**, but Docker “forgets”
that you stopped it and will start it again the next time the daemon (or
the host) starts. | Will start automatically. |
| `unless-stopped` | Container stays stopped **and** Docker remembers
that choice. | Will **not** start—because you had stopped it. |